+++ weblate/trans/templatetags/translations.py
@@ -17,7 +17,6 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 
-
 import re
 from datetime import date
 from uuid import uuid4
@@ -31,6 +30,7 @@ from django.utils.encoding import force_str
 from django.utils.html import escape
 from django.utils.safestring import mark_safe
 from django.utils.translation import gettext, gettext_lazy, ngettext, pgettext
+from siphashc import siphash
 
 from weblate.accounts.avatar import get_user_display
 from weblate.accounts.models import Profile
@@ -50,6 +50,7 @@ from weblate.trans.models.translation import GhostTranslation
 from weblate.trans.simplediff import html_diff
 from weblate.trans.util import get_state_css, split_plural
 from weblate.utils.docs import get_doc_url
+from weblate.utils.hash import hash_to_checksum
 from weblate.utils.markdown import render_markdown
 from weblate.utils.stats import BaseStats, ProjectLanguageStats
 
@@ -801,3 +802,9 @@ def percent_format(number):
     return pgettext("Translated percents", "%(percent)s%%") % {
         "percent": intcomma(int(number))
     }
+
+
+@register.filter
+def hash_text(name):
+    """Hash text for use in HTML id."""
+    return hash_to_checksum(siphash("Weblate URL hash", name.encode()))