+++ docs/changes.rst
@@ -31,6 +31,7 @@ Not yet released.
 * Show translations in user language first in the list.
 * Renamed shapings to more generic name variants.
 * Added new quality checks: :ref:`check-unnamed-format`, :ref:`check-long-untranslated`, :ref:`check-duplicate`.
+* Reintroduced support for wiping translation memory.
 
 Weblate 4.0.4
 -------------